+++ lib/ts/recipe/emailverification/recipe.ts
@@ -14,6 +14,7 @@
  */
 
 import { CreateRecipeFunction, NormalisedAppInfo } from "../../types";
+import { SessionClaimValidatorStore } from "supertokens-website/utils/sessionClaimValidatorStore";
 import RecipeModule from "../recipeModule";
 import { InputType, NormalisedInputType, PreAndPostAPIHookAction, RecipeInterface } from "./types";
 import { normaliseUserInput } from "./utils";
@@ -21,10 +22,15 @@ import RecipeImplementation from "./recipeImplementation";
 import OverrideableBuilder from "supertokens-js-override";
 import { checkForSSRErrorAndAppendIfNeeded, isTest } from "../../utils";
 import { UserInput } from "./types";
+import { EmailVerificationClaimClass } from "./emailVerificationClaim";
+import { PostSuperTokensInitCallbacks } from "../../postSuperTokensInitCallbacks";
 
 export default class Recipe implements RecipeModule<PreAndPostAPIHookAction, NormalisedInputType> {
     static instance?: Recipe;
     static RECIPE_ID = "emailverification";
+    static EmailVerificationClaim = new EmailVerificationClaimClass(
+        () => Recipe.getInstanceOrThrow().recipeImplementation
+    );
 
     config: NormalisedInputType;
     recipeImplementation: RecipeInterface;
@@ -40,6 +46,12 @@ export default class Recipe implements RecipeModule<PreAndPostAPIHookAction, Nor
             })
         );
         this.recipeImplementation = builder.override(this.config.override.functions).build();
+
+        PostSuperTokensInitCallbacks.addPostInitCallback(() => {
+            SessionClaimValidatorStore.addClaimValidatorFromOtherRecipe(
+                Recipe.EmailVerificationClaim.validators.isVerified(10)
+            );
+        });
     }
 
     static init(config?: UserInput): CreateRecipeFunction<PreAndPostAPIHookAction> {