+++ lib/ts/recipe/passwordless/index.ts
@@ -109,7 +109,7 @@ export default class RecipeWrapper {
      *
      * @param options (OPTIONAL) Use this to configure additional properties (for example pre api hooks)
      *
-     * @returns `{status: "OK", user, createdUser: bool}` if succesful
+     * @returns `{status: "OK", user, createdNewUser: bool}` if succesful
      *
      * @returns `{status: "INCORRECT_USER_INPUT_CODE_ERROR", failedCodeInputAttemptCount, maximumCodeInputAttempts}` if the code is incorrect
      *
@@ -120,7 +120,7 @@ export default class RecipeWrapper {
      * @throws STGeneralError if the API exposed by the backend SDKs returns `status: "GENERAL_ERROR"`
      */
     static async consumeCode(
-        input:
+        input?:
             | {
                   userInputCode: string;
                   userContext?: any;
@@ -133,7 +133,7 @@ export default class RecipeWrapper {
     ): Promise<
         | {
               status: "OK";
-              createdUser: boolean;
+              createdNewUser: boolean;
               user: PasswordlessUser;
               fetchResponse: Response;
           }