+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -12,7 +12,10 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-import { PreAndPostAPIHookAction as EmailPasswordPreAndPostAPIHookAction } from "../emailpassword/types";
+import {
+    UserType as EmailPasswordUserType,
+    PreAndPostAPIHookAction as EmailPasswordPreAndPostAPIHookAction,
+} from "../emailpassword/types";
 import {
     RecipePostAPIHookContext,
     RecipePreAPIHookContext,
@@ -21,12 +24,11 @@ import {
 import { PreAndPostAPIHookAction as ThirdPartyPreAndPostAPIHookAction, StateObject } from "../thirdparty/types";
 import { RecipeFunctionOptions } from "../recipeModule/types";
 import {
-    UserType as EmailPasswordUserType,
     InputType as AuthRecipeInputType,
     NormalisedInputType as AuthRecipeNormalisedInputType,
-} from "../authRecipeWithEmailVerification/types";
+} from "../authRecipe/types";
+
 import OverrideableBuilder from "supertokens-js-override";
-import { InputTypeOverride as EmailVerificationOverride } from "../emailverification/types";
 import { ThirdPartyUserType } from "../thirdparty/types";
 
 export type PreAndPostAPIHookAction = EmailPasswordPreAndPostAPIHookAction | ThirdPartyPreAndPostAPIHookAction;
@@ -39,7 +41,6 @@ export type UserInput = {
      * Refer to {@link https://supertokens.com/docs/thirdpartyemailpassword/advanced-customizations/frontend-functions-override/about the documentation}
      */
     override?: {
-        emailVerification?: EmailVerificationOverride;
         functions?: (
             originalImplementation: RecipeInterface,
             builder: OverrideableBuilder<RecipeInterface>